+++ .pipelines/pip-scripts/test-pip-wheels.sh
@@ -4,26 +4,35 @@ PYTHON_VERSION=${1:-3.11}
 
 export DEBIAN_FRONTEND=noninteractive
 
-# Try to prevent stochastic segfault from libc-bin
-echo "Reinstalling libc-bin..."
-rm /var/lib/dpkg/info/libc-bin.*
-apt-get clean
-apt-get update
-apt install -q libc-bin
+if [ "$MAC_BUILD" == "OFF" ]; then
+    # Try to prevent stochastic segfault from libc-bin
+    echo "Reinstalling libc-bin..."
+    rm /var/lib/dpkg/info/libc-bin.*
+    apt-get clean
+    apt-get update
+    apt install -q libc-bin
 
-# Update and install dependencies needed for testing
-echo "Installing apt dependencies..."
-apt-get update
-apt-get install -q -y \
-    python3 python3-pip python3-venv python3-dev \
-    libssl-dev zlib1g-dev libbz2-dev libreadline-dev libsqlite3-dev \
-    libncurses5-dev xz-utils tk-dev libxml2-dev libxmlsec1-dev \
-    libffi-dev liblzma-dev \
-    libopenblas-dev \
-    libboost-all-dev \
-    wget \
-    curl \
-    unzip
+    # Update and install dependencies needed for testing
+    echo "Installing apt dependencies..."
+    apt-get update
+    apt-get install -q -y \
+        python3 python3-pip python3-venv python3-dev \
+        libssl-dev zlib1g-dev libbz2-dev libreadline-dev libsqlite3-dev \
+        libncurses5-dev xz-utils tk-dev libxml2-dev libxmlsec1-dev \
+        libffi-dev liblzma-dev \
+        libopenblas-dev \
+        libboost-all-dev \
+        wget \
+        curl \
+        unzip
+elif [ "$MAC_BUILD" == "ON" ]; then
+    brew update
+    brew upgrade
+    brew install \
+        wget \
+        curl \
+        unzip
+fi
 
 # Install pyenv to use non-system python3 versions
 export PYENV_ROOT="/workspace/.pyenv"
